test(github-client): request/エラー処理系テストを別ファイルへ分離
【背景】直前コミットでlifecycleテストを分離したのに続き、test_github_client.pyの
肥大化を解消するため、HTTPリクエスト・リトライ・タイムアウト・sanitization関連の
テストを専用ファイルへ移動する。

【変更内容】リトライ動作、タイムアウト種別ごとのハンドリング、レート制限、
機密情報のsanitizationなど22件のテスト関数をtest_github_client_request.pyへ移動。
テストの実装内容・アサーションは変更なし（AST等価な移動のみ）。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
